requirements.txt
requests
beautifulsoup4
schedule
python-dotenv
fastapi
uvicorn[standard]
jinja2
python-multipart
